add protostar_project_section_test
